#9 refactor: 엔터티 변동에 따른 import문과 변수명 수정
기능 수정은 없음
