+++ src/main/java/com/backoffice/upjuyanolja/domain/room/repository/RoomImageRepository.java
@@ -1,6 +1,6 @@
 package com.backoffice.upjuyanolja.domain.room.repository;
 
-import com.backoffice.upjuyanolja.domain.room.entity.RoomImage;
+import com.backoffice.upjuyanolja.domain.accommodation.entity.RoomImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RoomImageRepository extends JpaRepository<RoomImage, Long> {
